+++ pricing/src/epilot/availability_api.py
@@ -13,6 +13,7 @@ class AvailabilityAPI:
         self.sdk_configuration = sdk_config
         
     
+    
     def dollar_availability_check(self, availability_check_params: shared.AvailabilityCheckParams, x_ivy_org_id: str) -> operations.DollarAvailabilityCheckResponse:
         r"""availabilityCheck
         The availability check endpoint
@@ -34,7 +35,10 @@ class AvailabilityAPI:
         headers['Accept'] = 'application/json'
         headers['user-agent'] = self.sdk_configuration.user_agent
         
-        client = self.sdk_configuration.security_client
+        if callable(self.sdk_configuration.security):
+            client = utils.configure_security_client(self.sdk_configuration.client, self.sdk_configuration.security())
+        else:
+            client = utils.configure_security_client(self.sdk_configuration.client, self.sdk_configuration.security)
         
         http_res = client.request('POST', url, data=data, files=form, headers=headers)
         content_type = http_res.headers.get('Content-Type')