+++ pricing/src/epilot/catalog_api.py
@@ -13,6 +13,7 @@ class CatalogAPI:
         self.sdk_configuration = sdk_config
         
     
+    
     def dollar_search_catalog(self, catalog_search: shared.CatalogSearch, x_ivy_org_id: str) -> operations.DollarSearchCatalogResponse:
         r"""searchCatalog
         Provides a querying functionalities over products and prices of the Catalog for a given organization.
@@ -34,7 +35,10 @@ class CatalogAPI:
         headers['Accept'] = 'application/json'
         headers['user-agent'] = self.sdk_configuration.user_agent
         
-        client = self.sdk_configuration.security_client
+        if callable(self.sdk_configuration.security):
+            client = utils.configure_security_client(self.sdk_configuration.client, self.sdk_configuration.security())
+        else:
+            client = utils.configure_security_client(self.sdk_configuration.client, self.sdk_configuration.security)
         
         http_res = client.request('POST', url, data=data, files=form, headers=headers)
         content_type = http_res.headers.get('Content-Type')