+++ pricing/src/epilot/models/shared/taxamount.py
@@ -8,11 +8,6 @@ from epilot import utils
 from typing import Optional, Union
 
 
-@dataclasses.dataclass
-class TaxAmountTax:
-    pass
-
-
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class TaxAmount: