+++ pricing/src/epilot/models/shared/totaldetails.py
@@ -9,16 +9,6 @@ from epilot import utils
 from typing import List, Optional, Union
 
 
-@dataclasses.dataclass
-class TotalDetailsRecurrences:
-    pass
-
-
-@dataclasses.dataclass
-class TotalDetailsTaxes:
-    pass
-
-
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class Breakdown: