+++ pricing/src/epilot/order_api.py
@@ -16,6 +16,7 @@ class OrderAPI:
         self.sdk_configuration = sdk_config
         
     
+    
     def create_order(self, request: shared.OrderPayload) -> operations.CreateOrderResponse:
         r"""createOrder
         Create an order
@@ -32,7 +33,10 @@ class OrderAPI:
         headers['Accept'] = 'application/json'
         headers['user-agent'] = self.sdk_configuration.user_agent
         
-        client = self.sdk_configuration.security_client
+        if callable(self.sdk_configuration.security):
+            client = utils.configure_security_client(self.sdk_configuration.client, self.sdk_configuration.security())
+        else:
+            client = utils.configure_security_client(self.sdk_configuration.client, self.sdk_configuration.security)
         
         http_res = client.request('POST', url, data=data, files=form, headers=headers)
         content_type = http_res.headers.get('Content-Type')
@@ -58,6 +62,7 @@ class OrderAPI:
         return res
 
     
+    
     def put_order(self, order_payload: shared.OrderPayload, id: str) -> operations.PutOrderResponse:
         r"""putOrder
         Update an existing Order
@@ -79,7 +84,10 @@ class OrderAPI:
         headers['Accept'] = 'application/json'
         headers['user-agent'] = self.sdk_configuration.user_agent
         
-        client = self.sdk_configuration.security_client
+        if callable(self.sdk_configuration.security):
+            client = utils.configure_security_client(self.sdk_configuration.client, self.sdk_configuration.security())
+        else:
+            client = utils.configure_security_client(self.sdk_configuration.client, self.sdk_configuration.security)
         
         http_res = client.request('PUT', url, data=data, files=form, headers=headers)
         content_type = http_res.headers.get('Content-Type')